rekognition/requirements.txt
boto3
pymongo
Pillow
requests
numpy